Automation and generation of mzML demo data
The goal is to generate small mzML files, compressed in .gz, containing only the time range and m/z range of interest so that files can be read and processed rapidly by OligoR, and shared easily (smaller size).

Creation of a template Excel mzML convert generator.xlsx file where the user can fill in the tube number (which is present in the raw data filenames), start and end TIC time to keep in the filtered file, and m/z window to keep as well.

The Excel file converts the time range from minutes to seconds for compatibility with msconvert, and generates a file number from 1 to n, where n is the number of files.

Creation of an R script to generate a single msconvert command line generating all filtered, converted and compressed data files.

The user must indicate the location of the demo data time filter files as well as input and output file templates.
The final line returns a single command line that can be pasted and executed in a terminal (requires the installation of msconvert)